requirements.txt
streamlit==1.11.0
pandas==1.3.4
inflection==0.5.1
plotly==5.13.0
folium==0.14.0
streamlit-folium==0.11.0
Pillow==8.4.0